User Location and Add Gem changes
The app now queries the user location occasionally to tell how far the
user is from a gem. This way, we can tell the user how far a gem is,
and only display gems in the area of a mile. Also edited was the
add_gem UI to be radio buttons instead of check boxes.